Add details to image collections automagically
If a user views an image from an image collection and then creates
an image detail, then the detail will automatically get added to the
same collection in addition to 'My Image Details'. If the user views
the image from anywhere else (from a search or another collection,
etc.) the detail will *only* appear in 'My Image Details'.